+++ src/components/ResultModal/style.scss
@@ -1,0 +1,95 @@
+.modal-inner {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin: 0 auto;
+  padding: 2rem;
+  max-width: 35rem;
+  min-width: 25rem;
+  min-height: 20rem;
+  border-radius: 0.8rem;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  box-shadow: 2px 2px 10px 4px #d2d2d2;
+  background: white;
+  z-index: 3;
+
+  .modal-header {
+    font-size: 2rem;
+    font-weight: bold;
+    margin: 0.6rem;
+  }
+
+  .modal-sub-header {
+    font-size: 1.3rem;
+    font-weight: bold;
+    color: #a0a0a0;
+    margin: 0.6rem 0;
+  }
+
+  .plus-icon {
+    margin: 0.5rem;
+  }
+
+  .modal-top-spacing {
+    position: absolute;
+    width: 100%;
+    height: 0.8rem;
+    background-color: rgb(71, 71, 71);
+    top: 0;
+    left: 0;
+    border-radius: 0.8rem 0.8rem 0 0;
+  }
+
+  .modal-close-button {
+    display: flex;
+    text-align: center;
+    justify-content: center;
+    width: 1.6rem;
+    height: 1.6rem;
+    position: absolute;
+    right: 1.5rem;
+    top: 1.5rem;
+    background-color: white;
+  }
+
+  .result-numbers-container {
+    display: flex;
+    width: 100%;
+    border-width: 0.5rem 0 0.5rem 0;
+    border-style: solid;
+    border-color: rgba(0, 0, 0, 0.1);
+    flex-direction: row;
+    padding: 1rem 0 1rem 0;
+    align-items: center;
+    justify-content: center;
+  }
+
+  .modal-numbers-container {
+    display: flex;
+    max-height: 25rem;
+    overflow: auto;
+    width: 100%;
+    padding: 0.5rem 0 0.5rem 0;
+    flex-direction: column;
+    &::-webkit-scrollbar {
+      display: none;
+    }
+  }
+
+  .modal-result-text {
+    display: flex;
+    width: 100%;
+    flex-direction: column;
+    margin: 1rem 0 1rem 0;
+    border-width: 0.2rem 0 0.2rem 0;
+    border-style: dashed;
+    border-color: rgba(0, 0, 0, 0.5);
+    padding: 1rem 0 1rem 0;
+    align-items: center;
+    gap: 0.6rem;
+  }
+}
