+++ src/components/UtilComponent/Modal/style.scss
@@ -1,0 +1,11 @@
+.modal {
+  font-family: 'Noto Sans KR', sans-serif;
+  display: flex;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  backdrop-filter: blur(5px);
+  z-index: 15;
+}
